Desktop/project/Resume-Parser-OpenAI-main/requirement.txt
flask
pypdf
openai>=1.21.0
pyyaml
